break down client into modular architecture
- Split monolithic client.py into focused modules:
  - configuration.py: ClientConfiguration and RateLimitConfig management
  - rate_limiter.py: Rate limiting with context manager support
  - prompt_builder.py: Batch and single prompt construction strategies
  - error_handler.py: Content-aware error handling and messages
  - content_processor.py: Orchestration between files subsystem and API
- Create new gemini_client.py as main orchestrator using composition
- Maintain clean separation: files handle extraction, client handles API
